+++ star-manager/src/main/java/com/rookiestar/starmanager/controller/ManagerRestController.java
@@ -4,17 +4,17 @@ import com.rookiestar.starmanager.constant.AttributeNames;
 import com.rookiestar.starmanager.constant.PermissionNames;
 import com.rookiestar.starmanager.constant.RoleNames;
 import com.rookiestar.starmanager.entity.company.Company;
+import com.rookiestar.starmanager.exception.RequestParameterException;
+import com.rookiestar.starmanager.rabbit.MessageProducer;
 import com.rookiestar.starmanager.entity.company.CompanyToReview;
 import com.rookiestar.starmanager.entity.department.Department;
 import com.rookiestar.starmanager.entity.employee.Employee;
 import com.rookiestar.starmanager.entity.manager.Manager;
 import com.rookiestar.starmanager.entity.position.Position;
 import com.rookiestar.starmanager.service.*;
-import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.annotation.Logical;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.apache.shiro.authz.annotation.RequiresRoles;
-import org.apache.shiro.session.Session;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -36,7 +36,7 @@ public class ManagerRestController {
     @Autowired
     private RetrieveService retrieveService;
     @Autowired
-    private EmailService emailService;
+    private MessageProducer messageProducer;
 
 
 
@@ -48,13 +48,23 @@ public class ManagerRestController {
      */
     @RequiresRoles(value = {RoleNames.MANAGER})
     @RequiresPermissions(value = {PermissionNames.WRITE})
-    @RequestMapping(value = "confirmCompanyRegisterApply.do")
-    public Company confirmCompanyRegisterApply(int companyId, String name, String legalRepresentativeName, String email, String address, String phone){
+    @RequestMapping(value = "/confirmCompanyRegisterApply.do")
+    public Company confirmCompanyRegisterApply(Integer companyId, String name, String legalRepresentativeName, String email, String address, String phone){
+        if(companyId==null||name==null||legalRepresentativeName==null||email==null||address==null||phone==null){
+            throw new RequestParameterException("请求参数不正确");
+        }
         deleteService.deleteCompanyToReviewByCompanyId(companyId);
         Company company=new Company(name,legalRepresentativeName,email,address,phone);
         Company newCompany=createService.registerCompany(company);
         String content = "您的公司注册申请已被管理员通过，您的公司ID为："+company.getCompanyId();
-        emailService.sendSimpleEmail("2019302110260@whu.edu.cn","注册成功通知",content);
+
+        Map<String,String> contentMap = new HashMap<>(10);
+        contentMap.put("to",email);
+        contentMap.put("subject","注册成功通知");
+        contentMap.put("content",content);
+
+        messageProducer.sendNotice(contentMap);
+
         return newCompany;
     }
